+++ sdk/app/theme_test.go
@@ -1,0 +1,15 @@
+package app_test
+
+import (
+	"testing"
+
+	"github.com/wrzfeijianshen/fyne2/sdk/app"
+	"github.com/wrzfeijianshen/fyne2/test"
+)
+
+func TestApplySettings_BeforeContentSet(t *testing.T) {
+	a := test.NewApp()
+	_ = a.NewWindow("NoContent")
+
+	app.ApplySettings(a.Settings(), a)
+}